% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg.R
\name{autoupdate.cfg}
\alias{autoupdate.cfg}
\title{Auto-Update Configuration}
\usage{
autoupdate.cfg(path = openbeta_server.path())
}
\arguments{
\item{path}{Path of DCS World installation}
}
\value{
List of character and logical vectors; see examples
}
\description{
Auto-Update Configuration
}
\details{
Reads \code{autoupdate.cfg} belonging to the OpenBeta Server by
default.
}
\examples{
dcsworldr::autoupdate.cfg() \%>\% utils::str()
#> List of 10
#>  $ WARNING       : chr "DO NOT EDIT this file. You may break your install!"
#>  $ branch        : chr "openbeta"
#>  $ version       : chr "2.7.10.19473"
#>  $ timestamp     : chr "20220207-145447"
#>  $ arch          : chr "x86_64"
#>  $ lang          : chr "EN"
#>  $ modules       : chr "DEDICATED_SERVER"
#>  $ launch        : chr "bin/DCS.exe"
#>  $ launch_args   : chr [1:2] "--norender" "--server"
#>  $ no_auto_inject: logi TRUE
}
