% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg.R
\name{cfg.is_dedicated_server}
\alias{cfg.is_dedicated_server}
\title{Auto-Update Configuration for Dedicated Server}
\usage{
cfg.is_dedicated_server(cfg = autoupdate.cfg())
}
\arguments{
\item{cfg}{List of auto-update configuration pairs}
}
\value{
Logical \code{TRUE} if modules includes dedicated server
}
\description{
Auto-Update Configuration for Dedicated Server
}
\examples{
dcsworldr::cfg.is_dedicated_server()
#> [1] TRUE
}
