% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsworldr-package.R
\docType{package}
\name{dcsworldr-package}
\alias{dcsworldr}
\alias{dcsworldr-package}
\title{dcsworldr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: R3D L1N3 \email{r0y@digitalcombatsimulator.world} (\href{https://orcid.org/0000-0001-6594-1280}{ORCID})

}
\keyword{internal}
