% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedicated-server.R
\name{dedicated_server.is_installed}
\alias{dedicated_server.is_installed}
\title{Dedicated Server is Installed}
\usage{
dedicated_server.is_installed()
}
\value{
logical; \code{TRUE} if installed and configured as a dedicated
server, \code{FALSE} if not installed or not installed with the dedicated
server module
}
\description{
Success depends on multiple factors:
\itemize{
\item the existence of the correctly installed folders; and
\item the presence and correct formatting of the installation's
\code{autoupdate.cfg} JSON file;
\item the modules array must include the token \code{DEDICATED_SERVER}.
}
}
