% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{eagle_dynamics.path}
\alias{eagle_dynamics.path}
\title{Path of Eagle Dynamics}
\usage{
eagle_dynamics.path(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[base:file.path]{base::file.path}}
  \describe{
    \item{\code{fsep}}{the path separator to use (assumed to be ASCII).}
  }}
}
\value{
Existing directory path; fails if Eagle Dynamics with Program Files
does not exist
}
\description{
Fails if \emph{no} Eagle Dynamics simulators or other products have not yet
been installed.
}
\details{
Any addition dot parameters append to Eagle Dynamics as sub-path components
\emph{prior} to checking for directory existence. Hence any sub-path must
address a directory and not a file.
}
