% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{openbeta_server.path}
\alias{openbeta_server.path}
\title{Path of OpenBeta Server}
\usage{
openbeta_server.path(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=eagle_dynamics.path]{eagle_dynamics.path}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Existing directory path
}
\description{
Fails if the OpenBeta Server product has not been installed.
}
\examples{
\dontrun{
library(dcsworldr)

# List all files with extension `.miz` found recursively under the OpenBeta
# Server folder. Escape the full stop and match the end of the file name.
# Name the vector of paths by the base name.
file.miz <- openbeta_server.path() \%>\%
  list.files(pattern = "\\\\.miz$", full.names = TRUE, recursive = TRUE)
names(file.miz) <- basename(file.miz)

# Are any duplicated by base name? Not at the time of writing.
basename(file.miz) \%>\%
  duplicated() \%>\%
  any()
#> [1] FALSE
}
}
