% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{program_files.path}
\alias{program_files.path}
\title{Path of Program Files on Windows}
\usage{
program_files.path(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[base:file.path]{base::file.path}}
  \describe{
    \item{\code{fsep}}{the path separator to use (assumed to be ASCII).}
  }}
}
\value{
Existing directory path
}
\description{
Path of Program Files on Windows
}
